# src/main/java/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Finance App API","description":"List of all Endpoints of the project \n Except endpoints for connecting Bank","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Fake Connection","description":"Use this if u don't have opportunity to connect a real bank"},{"name":"Users","description":"Use this if u don't have opportunity to connect a real bank"},{"name":"Report","description":"Connected with report things"}],"paths":{"/report/save-comment/{year}/{month}":{"post":{"tags":["Report"],"operationId":"saveComments","parameters":[{"name":"month","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"year","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentData"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/report/get":{"post":{"tags":["Report"],"operationId":"getReport","parameters":[{"name":"month","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"year","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/report/createReport/startend":{"post":{"tags":["Report"],"operationId":"createReport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSpanData"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/report/createReport/monthYear":{"post":{"tags":["Report"],"operationId":"createReport_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MonthYear"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/report/createReport/auto":{"post":{"tags":["Report"],"operationId":"createReport_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/report/category-history":{"post":{"tags":["Report"],"operationId":"getHistory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryHistory"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/label/set":{"post":{"tags":["Label"],"summary":"Установить список лейблов для транзакции","operationId":"setLabelsForTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataForSet"}}},"required":true},"responses":{"200":{"description":"Лейблы устанавлены","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Ошибочка ","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppError"}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/goals/delete":{"post":{"tags":["Finance Goal"],"operationId":"deleteGoal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveData"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/goals/create":{"post":{"tags":["Finance Goal"],"operationId":"createGoal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataForGoal"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/fake/createAccounts":{"post":{"tags":["Fake Connection"],"operationId":"createAccounts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}},"security":[{"javainuseapi":[]}]}},"/fake/createAccount":{"post":{"tags":["Fake Connection"],"operationId":"createAccount","parameters":[{"name":"Type","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}},"security":[{"javainuseapi":[]}]}},"/fake/addTransaction":{"post":{"tags":["Fake Connection"],"operationId":"addTransaction","parameters":[{"name":"IdAccount","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AbstractTransaction"}}}}}},"security":[{"javainuseapi":[]}]}},"/category/set":{"post":{"tags":["Categories"],"operationId":"setCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetCategoryData"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}},"/api/user/register":{"post":{"tags":["Users"],"operationId":"createPerson","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataForRegister"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWrapper"}}}}}}},"/api/user/checkEmail":{"post":{"tags":["Users"],"operationId":"checkEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailCheck"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/fake/getAll":{"get":{"tags":["Fake Connection"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}},"security":[{"javainuseapi":[]}]}},"/acc/sync-all":{"get":{"tags":["Bank Connection"],"summary":"Синронизировать все счета","operationId":"SyncAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"javainuseapi":[]}]}}},"components":{"schemas":{"CommentData":{"type":"object","properties":{"comment":{"type":"string"}}},"TimeSpanData":{"type":"object","properties":{"start":{"type":"string"},"end":{"type":"string"},"formatStart":{"type":"string","format":"date"},"formatEnd":{"type":"string","format":"date"},"timeEnd":{"type":"string","format":"date-time"},"timeStart":{"type":"string","format":"date-time"}}},"MonthYear":{"type":"object","properties":{"month":{"type":"integer","format":"int32"},"year":{"type":"integer","format":"int32"}}},"CategoryHistory":{"type":"object","properties":{"categoryId":{"type":"integer","format":"int64"},"length":{"type":"integer","format":"int32"}}},"DataForSet":{"type":"object","properties":{"transactionId":{"type":"integer","format":"int64"},"labelsIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"AppError":{"type":"object","properties":{"error":{"type":"string"}}},"RemoveData":{"type":"object","properties":{"goalId":{"type":"integer","format":"int64"}}},"DataForGoal":{"type":"object","properties":{"friendlyName":{"type":"string"},"goalAmount":{"type":"number","format":"double"},"date":{"type":"string"},"accountId":{"type":"string","format":"uuid"},"url":{"type":"string"}}},"AbstractTransaction":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"amount":{"type":"number","format":"double"},"description":{"type":"string"},"dateTime":{"type":"string","format":"date-time"},"category":{"$ref":"#/components/schemas/Category"},"merchant":{"type":"string"},"labelList":{"type":"array","items":{"$ref":"#/components/schemas/Label"}},"account":{"$ref":"#/components/schemas/Account"},"labelsIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"Account":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"lastSync":{"type":"string","format":"date-time"},"balance":{"type":"number","format":"double"},"idInSystem":{"type":"string"},"friendlyName":{"type":"string"},"provider":{"type":"string"},"user":{"$ref":"#/components/schemas/CustomUser"},"transactionList":{"type":"array","items":{"$ref":"#/components/schemas/AbstractTransaction"}},"name":{"type":"string"},"imageProviderUrl":{"type":"string"}}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"categoryName":{"type":"string"}}},"CustomUser":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"role":{"type":"string"},"accountList":{"type":"array","items":{"$ref":"#/components/schemas/Account"}},"email":{"type":"string"},"password":{"type":"string"},"enabled":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"username":{"type":"string"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Label":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"SetCategoryData":{"type":"object","properties":{"transactionId":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"}}},"DataForRegister":{"type":"object","properties":{"password":{"type":"string"},"email":{"type":"string"}}},"UserWrapper":{"type":"object","properties":{"username":{"type":"string"}}},"EmailCheck":{"type":"object","properties":{"email":{"type":"string"}}}},"securitySchemes":{"javainuseapi":{"type":"http","in":"header","scheme":"basic"}}}}